+++ Net_Core/Web.Angular/projects/gw-lib/core/pick-list/src/c/list/list.component.ts
@@ -5,6 +5,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 // Library
 import { ModalOptions, ModalService, ModalSize } from '@growthware/core/modal';
+import { encapsulateStyle } from '@angular/compiler';
 
 @Component({
 	selector: 'gw-core-list',
@@ -26,7 +27,10 @@ export class ListComponent implements OnInit {
 	allItemsText = input<string>('');
 	availableItems = input<Array<string>>([]);
 	availableItemsChange = output<Array<string>>();
+	pickListTableContentsBackground = input<string>('pink');
+	fontColor = input<string>('black');
 	header = input<string>('');
+	pickListTableHeaderBackground = input<string>('lightpink');
 	id = input<string>('');
 	name = input<string>('');
 	pickListTableHelp = input<string>('');
@@ -36,6 +40,8 @@ export class ListComponent implements OnInit {
 
 	ngOnInit(): void {
 		this._ModalOptions = new ModalOptions(this.id() + '_Modal', this.header(), this.pickListTableHelp(), ModalSize.Small);
+		document.documentElement.style.setProperty('--pickListTableContentsBackground', this.pickListTableContentsBackground());
+		document.documentElement.style.setProperty('--pickListTableHeaderBackground', this.pickListTableHeaderBackground());
 	}
 
 	onShowHelp(): void {