+++ Release_Notes.md
@@ -16,6 +16,7 @@
 - Re-Worked the navagation feature to reduce the number of components needed
 - The Toast feature now works on signals only
 - The Dynamic Table now uses ClientChoices.subHeadColor to set the header font color
+- Added color properties to ListComponent
 
 ### Bug Fixes
 - Fixed NG0956: The configured tracking expression (track by identity) in file-list.component.html